\docType{methods}
\name{addRow}
\alias{addRow}
\alias{addRow,DXGTable-method}
\title{Add a Row to a GTable}
\arguments{
  \item{handler}{A data object handler}

  \item{row}{A list of row values, either in exact column
  order, or named with column names.}
}
\description{
  Takes in a list of values and adds it to the referenced
  GTable.
}
\seealso{
  \code{\link{addRows}} for adding multiple rows at once
  from a data frame
}

