\docType{methods}
\name{addRows}
\alias{addRows}
\alias{addRows,DXGTable-method}
\title{Add Rows to a GTable}
\arguments{
  \item{handler}{A data object handler}

  \item{rows}{A data frame containing the rows to add}
}
\description{
  Takes in a data frame and adds it as rows to the
  referenced GTable.
}
\seealso{
  \code{\link{addRow}} for adding single rows at a time
}

