% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_evaluation_sheet.R
\name{build_evaluation_sheet}
\alias{build_evaluation_sheet}
\title{Build Evaluation Sheet}
\usage{
build_evaluation_sheet(
  evaluation_name,
  evaluation_description = NA_character_,
  panel_data = NA,
  panel_data_filepath = NA_character_,
  analytes,
  targets,
  qualitative_outcomes = c("Positive", "Negative"),
  semiquantitative_outcomes = NA_character_,
  quantitative_units = NA_character_
)
}
\arguments{
\item{evaluation_name}{The name of the evaluation being used for the
evaluation. Must be a character string (not a multi-level vector) and must
not be \code{NA}.}

\item{evaluation_description}{An optional sentence or paragraph describing
the evaluation. Defaults to \code{NA}. If provided, it must be a string
rather than a multi-level vector.}

\item{panel_data}{The output of \code{build_panel_sheet}. Either this or a
panel_data_filepath must be provided, but not both. Defaults to \code{NA}.}

\item{panel_data_filepath}{An excel file generated by
\code{write_panel_sheet()}. Either this or panel_data mupst be provided,
but not both. Defaults to \code{NA}.}

\item{analytes}{The analyte or analytes to be included in the evaluation.
This must be provided with at least one level (not \code{NA}). This
function will stop with an error if the analytes to be evaluated arte not
included in the sample panel.}

\item{targets}{The target or targets for which the assay under evaluation.
For example, for a SARS-CoV-2 assay, this could be \code{"Spike"} or
\code{"RBD"}. Multiple targets can also be used, such as \code{c("Spike",
  "Nucleocapsid")}. This must be provided with at least one level (not
\code{NA}).}

\item{qualitative_outcomes}{The valid qualitative outcomes associated with
the evaluation. Typically, this should be \code{c("Positive", "Negative")}.
Note that while the assay being evaluated may also provide an "Equivocal"
result level, equivocal results will be counted against the assay in terms
of performance. That is, an equivocal result on a positive sample will be
called as a false negative result and an equivocal result on a negative
sample will be called as a false positive result.}

\item{semiquantitative_outcomes}{The valid semi-quantitative outcomes
associated with the evaluation. Defaults to \code{NA}.}

\item{quantitative_units}{If quantitative outcomes have been established for
the evaluation, this is a character string describing the units of those
quantitative results. Defaults to \code{NA}. If the results are unit-less,
\code{"Unit-less"} can be used.}
}
\value{
Returns a list including the \code{evaluation_metadata} and
\code{evaluation_table}
}
\description{
A function to build an excel file with fill-in-the-blank cells for common
information needed about a sample evaluation.
}
\examples{
build_evaluation_sheet(
  evaluation_name = "Example Evaluation",
  evaluation_description = NA_character_,
  panel_data =
    build_panel_sheet(
      panel_name = "Example Panel",
      panel_description = "An example panel.",
      n_samples = 5L,
      sample_groups = "Samples",
      sample_matrices = "Serum",
      analytes = c("IgM", "IgG", "Pan-Ig"),
      targets = "Spike",
      qualitative_outcomes = c("Positive", "Negative"),
      qualitative_comparators = "Authorized NAAT and CDC Assay",
      semiquantitative_outcomes = NA,
      semiquantitative_comparators = NA,
      quantitative_units = NA,
      quantitative_comparators = NA
    ),
  analytes = c("IgM", "IgG", "Pan-Ig"),
  targets = "Spike",
  qualitative_outcomes = c("Positive", "Negative"),
  semiquantitative_outcomes = NA_character_,
  quantitative_units = NA_character_
)
}
