% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_performance.R
\name{calculate_performance}
\alias{calculate_performance}
\title{Calculate Performance}
\usage{
calculate_performance(panel_data, evaluation_data)
}
\arguments{
\item{panel_data}{The result of \code{read_panel()}.}

\item{evaluation_data}{The result of \code{read_evaluation()}.}
}
\value{
A table of summary statistics.
}
\description{
\strong{UNDER DEVELOPMENT}
}
\details{
Generate a summary statistics table.
}
\examples{
calculate_performance(dxr::example_panel_1, dxr::example_evaluation_1)
}
