% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{confidence}
\alias{confidence}
\title{Confidence}
\usage{
confidence(true_calls, false_calls, interval = 0.95)
}
\arguments{
\item{true_calls}{The number of correct calls (true positives or true
negatives).}

\item{false_calls}{The number of incorrect calls (false positives or false
negatives).}

\item{interval}{The confidence interval to calculate. Defaults to
\code{0.95} for 95\%.}
}
\value{
A list including a \code{lower_bound} and an \code{upper_bound}.
}
\description{
Calculate a confidence interval for a proportion, such as PPA or NPA. Uses
the recommended method as described in Altman chapter 6 (loc 1179 in the
Kindle edition) and in CLSI EP12-A2 (2008). Note that the CLSI method may
round the z-score for a 95\% CI to 1.96, leading to slight differences in
computed values. Here, the z-score is calculated using the
\code{stats::qnorm()} function.
}
\details{
Also known as "Wilson's method".
}
\examples{
confidence(100, 0, 0.95)
confidence(100, 0)
confidence(100, 0, 0.99)
confidence(99, 1)
}
