% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_reactivity.R
\name{cross_reactive_subgroup}
\alias{cross_reactive_subgroup}
\title{Determine if a subgroup of negatives is cross-reactive}
\usage{
cross_reactive_subgroup(
  true_calls_overall,
  false_calls_overall,
  true_calls_subgroup,
  false_calls_subgroup,
  digits = 0.1,
  interval = 0.95
)
}
\arguments{
\item{true_calls_overall}{The number of true calls (true negatives) in all
samples tested.}

\item{false_calls_overall}{The number of false calls (false positives) in all
samples tested.}

\item{true_calls_subgroup}{The number of true calls (true negatives) in the
subgroup of interest (e.g. HIV+).}

\item{false_calls_subgroup}{The number of false calls (false positives) in
the subgroup of interest (e.g. HIV+).}

\item{digits}{The number of digits to round to. Defaults to \code{0.1}.}

\item{interval}{The confidence interval to calculate. Defaults to \code{0.95}
for 95\%.}
}
\value{
A list containing the estimated difference, the confidence interval,
a boolean attribute indicating whether or not a difference was detected at
the specified confidence (\code{difference_detected}), and various strings
and formatted percents that can be used to convey the information.
}
\description{
This is a wrapper around \code{proportion_difference()}.
}
\examples{
cross_reactive_subgroup(
  true_calls_overall = 79,
  false_calls_overall = 1,
  true_calls_subgroup = 9,
  false_calls_subgroup = 1,
  digits = 0.1,
  interval = 0.95
)
}
