% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crossed_outcomes}
\alias{crossed_outcomes}
\title{Crossed Outcomes}
\usage{
crossed_outcomes(analytes, qualitative_outcomes)
}
\arguments{
\item{analytes}{The analytes to include.}

\item{qualitative_outcomes}{The qualitative outcomes to include.}
}
\value{
A vector of all permutations of analytes and qualitative outcomes.
}
\description{
Come up with all the permutations of qualitative outcomes for each sample,
including all analytes.
}
\examples{
dxr:::crossed_outcomes(c("IgM", "IgG"), c("Positive", "Negative"))

}
