\docType{methods}
\name{desc}
\alias{desc}
\alias{desc,DXGTable-method}
\title{Get the cached describe from a DNAnexus handler}
\arguments{
  \item{handler}{A DNAnexus handler}
}
\value{
  description of the referenced object, if cached;
  otherwise NA
}
\description{
  Returns the description of the referenced data object, if
  cached.
}

