\docType{methods}
\name{describe}
\alias{describe}
\alias{describe,DXGTable-method}
\title{Describe the Data Object}
\arguments{
  \item{handler}{A data object handler}
}
\value{
  named list of the data object's describe hash
}
\description{
  Returns the data frame containing columns describing the
  data object.
}

