\docType{methods}
\name{dim}
\alias{dim}
\alias{dim,DXGTable-method}
\title{Dimensions of a GTable}
\arguments{
  \item{x}{A GTable handler}
}
\value{
  number of rows and columns in the GTable
}
\description{
  Returns a vector of the number of rows and columns (not
  including the row ID column) of the referenced GTable,
  respectively.
}

