\name{DXGTable-class}
\alias{DXGTable-class}
\title{GenomicTable (GTable) Handler}
\description{
  This class is an interface to a GenomicTable(GTable)
  stored on the DNAnexus platform.
}
\section{Slots}{
  \describe{ \item{\code{id}}{String ID of the GTable}
  \item{\code{project}}{String ID of the GTable's project}
  \item{\code{.desc}}{Cached copy from calling describe on
  the GTable} }
}

