% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_evaluation_sheet.R
\name{evaluation_sheet_csv_method}
\alias{evaluation_sheet_csv_method}
\title{Evaluation Sheet CSV Method}
\usage{
evaluation_sheet_csv_method(evaluation_sheet_data)
}
\arguments{
\item{evaluation_sheet_data}{The evaluation data, which are the output of
\code{dxr::build_evaluation_sheet()}.}
}
\value{
Nothing for now, just an error.
}
\description{
Evaluation Sheet CSV Method
}
