% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_evaluation_sheet.R
\name{evaluation_sheet_excel_method}
\alias{evaluation_sheet_excel_method}
\title{Evaluation Sheet Excel Method}
\usage{
evaluation_sheet_excel_method(evaluation_sheet_data)
}
\arguments{
\item{evaluation_sheet_data}{The evaluation data, which are the output of
\code{dxr::build_evaluation_sheet()}.}
}
\value{
A workbook object from \code{openxlsx}.
}
\description{
Build an excel workbook, using \code{openxlsx}.
}
