\docType{methods}
\name{flushRows}
\alias{flushRows}
\alias{flushRows,DXGTable-method}
\title{Flush Added Rows to a GTable}
\arguments{
  \item{handler}{A GTable handler}
}
\description{
  Takes any buffered rows to be added to the GTable and
  adds them to the GTable.
}
\seealso{
  \code{\link{closeObj}} which calls this function if
  necessary
}

