% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_big.R
\name{format_big}
\alias{format_big}
\title{Format big numbers}
\usage{
format_big(x)
}
\arguments{
\item{x}{The number to format.}
}
\value{
\code{x}, formatted with a comma for the big mark. Scientific
notation is disabled.
}
\description{
A wrapper around \code{format()}.
}
