% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_percent.R
\name{format_percent}
\alias{format_percent}
\title{Format a percent}
\usage{
format_percent(percent, digits = 0.1)
}
\arguments{
\item{percent}{The percentage to be rounded, as a decimal (e.g. 0.95 for
95\%).}

\item{digits}{The number of digits to round to. Defaults to \code{0.1}.}
}
\value{
A formatted percent.
}
\description{
Specifically, a percent based on a proportion that cannot exceed 1. Round
and format all values to the specified digits using the
\code{scales::percent} function, but replace outputs like 100.0\%" and
"100.00\%" with "100\%".
}
\examples{
format_percent(0.95)
format_percent(0.9521231251, 0.0001)
}
