% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_evaluation_metadata}
\alias{get_evaluation_metadata}
\title{Get Evaluation Metadata}
\usage{
get_evaluation_metadata(evaluation_data)
}
\arguments{
\item{evaluation_data}{the output of \code{build_evaluation_sheet()} or
\code{read_evaluation()}.}
}
\value{
The evaluation metadata as a list.
}
\description{
A helper function to convert evaluation metadata from a nested list tibble to
a list.
}
