% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lengthen}
\alias{lengthen}
\title{Lengthen}
\usage{
lengthen(vec, desired_length)
}
\arguments{
\item{vec}{The vector}

\item{desired_length}{The length of the final vector}
}
\value{
A the input vector with \code{NA} appended until it reaches
\code{desired_length}.
}
\description{
Pad a vector to the desired length by adding \code{NA}.
}
