\name{loadFromEnvironment}
\alias{loadFromEnvironment}
\title{(Re)Load DNAnexus Configuration Values}
\usage{
  loadFromEnvironment()
}
\description{
  Load (or reload) state from: 1) shell environment
  variables, 2) configuration options saved by dx (the
  DNAnexus command-line client), 3) hardcoded defaults
}

