% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{lr_pos}
\alias{lr_pos}
\title{Positive Likelihood Ratio}
\usage{
lr_pos(sensitivity_estimate, specificity_estimate, digits = 0.1)
}
\arguments{
\item{sensitivity_estimate}{The estimated sensitivity or PPA with full
precision (do not round this number before passing it to this function).}

\item{specificity_estimate}{The estimated specificity or NPA with full
precision (do not round this number before passing it to this function).}

\item{digits}{The number of digits to round to. Defaults to \code{0.1}.}
}
\value{
A positive Likelihood Ratio (LR+)
}
\description{
Positive Likelihood Ratio
}
\examples{
lr_pos(
  sensitivity_estimate = 30/30,
  specificity_estimate = 78/80,
  digits = 0.01
)
}
