\docType{methods}
\name{ncol}
\alias{ncol}
\alias{ncol,DXGTable-method}
\title{The Number of Columns in the GTable}
\arguments{
  \item{x}{A GTable handler}
}
\value{
  number of columns in the GTable
}
\description{
  Returns the number of columns (not including the row ID
  column) in the GTable.
}

