% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_value.R
\name{npv}
\alias{npv}
\title{Negative Predictive Value}
\usage{
npv(sensitivity, specificity, prevalence)
}
\arguments{
\item{sensitivity}{The estimate of Sensitivity (PPA)}

\item{specificity}{The estimate of Specificity (NPA)}

\item{prevalence}{The estimate of the prevalence of the condition in the
population.}
}
\value{
The negative predictive value
}
\description{
Calculate Negative Predictive Value (NPV) in a population based on
sensitivity, specificity, and prevalence estimates.
}
\examples{
npv(0.95, 0.95, 0.05)
}
