% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_panel_sheet.R
\name{panel_sheet_csv_method}
\alias{panel_sheet_csv_method}
\title{Panel Sheet CSV Method}
\usage{
panel_sheet_csv_method(panel_sheet_data)
}
\arguments{
\item{panel_sheet_data}{The panel data, which are the output of
\code{dxr::build_panel_sheet()}.}
}
\value{
Nothing for now, just an error.
}
\description{
Panel Sheet CSV Method
}
