% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_panel_sheet.R
\name{panel_sheet_excel_method}
\alias{panel_sheet_excel_method}
\title{Panel Sheet Excel Method}
\usage{
panel_sheet_excel_method(panel_sheet_data)
}
\arguments{
\item{panel_sheet_data}{The panel data, which are the output of
\code{dxr::build_panel_sheet()}.}
}
\value{
A workbook object from \code{openxlsx}.
}
\description{
Build an excel workbook, using \code{openxlsx}.
}
