% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_value.R
\name{ppv}
\alias{ppv}
\title{Positive Predictive Value}
\usage{
ppv(sensitivity, specificity, prevalence)
}
\arguments{
\item{sensitivity}{The estimate of Sensitivity (PPA)}

\item{specificity}{The estimate of Specificity (NPA)}

\item{prevalence}{The estimate of the prevalence of the condition in the
population.}
}
\value{
The positive predictive value
}
\description{
Calculate Positive Predictive Value (PPV) in a population based on
sensitivity, specificity, and prevalence estimates.
}
\examples{
ppv(0.95, 0.95, 0.05)
}
