% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_value.R
\name{ppv_with_confidence}
\alias{ppv_with_confidence}
\title{Positive Predictive Value (with confidence interval)}
\usage{
ppv_with_confidence(sensitivity, specificity, prevalence, interval)
}
\arguments{
\item{sensitivity}{Output of a function such as \code{sensitivity()}.}

\item{specificity}{Output of a function such as \code{specificity()}.}

\item{prevalence}{The estimate of the prevalence of the condition in the
population.}

\item{interval}{The confidence interval used in the sensitivity and
specificity calculations.}
}
\value{
The positive predictive value, with confidence.
}
\description{
Calculate the positive predictive value (PPV) with a confidence interval
based on a sensitivity output from \code{agreement()},
\code{agreement_fraction()}, \code{sensitivity()}, or
\code{sensitivity_fraction()} and a specificity output from
\code{agreement()}, \code{agreement_fraction()}, \code{specificity()}, or
\code{specificity_fraction()}.
}
\examples{
ppa <- sensitivity(99, 0)
npa <- specificity(99, 0)
ppv_with_confidence(ppa, npa, 0.05, 0.95)
}
