\name{printenv}
\alias{printenv}
\title{Print DNAnexus Configuration Values}
\usage{
  printenv()
}
\description{
  Prints the current set of configuration values that are
  being used to contact the DNAnexus platform.
}
\examples{
# Running the following
printenv()

# Results in:
# Currently loaded environment:
#   DX_APISERVER_HOST: api.dnanexus.com
#   DX_APISERVER_PORT: 443
#   DX_APISERVER_PROTOCOL: https
}

