% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_difference.R
\name{proportion_difference}
\alias{proportion_difference}
\title{Calculate the confidence interval of a difference in proportions (e.g.
sensitivity or specificity)}
\usage{
proportion_difference(
  true_calls_1,
  false_calls_1,
  true_calls_2,
  false_calls_2,
  interval = 0.95,
  digits = 0.1
)
}
\arguments{
\item{true_calls_1}{The number of correct calls (true positives or true
negatives) in the first proportion of interest.}

\item{false_calls_1}{The number of incorrect calls (false positives or false
negatives) in the first proportion of interest.}

\item{true_calls_2}{The number of correct calls (true positives or true
negatives) in the second proportion of interest.}

\item{false_calls_2}{The number of incorrect calls (false positives or false
negatives in the second proportion of interest).}

\item{interval}{The confidence interval to calculate. Defaults to
\code{0.95} for 95\%.}

\item{digits}{The number of digits to round to. Defaults to \code{0.1}.}
}
\value{
A list containing the estimated difference, the confidence interval,
a boolean attribute indicating whether or not a difference was detected at
the specified confidence (\code{difference_detected}), and various strings
and formatted percents that can be used to convey the information.
}
\description{
Calculate a confidence interval for a difference in proportions, such as PPA
or NPA. Uses the recommended method as described in Altman chapter 6 (loc
1249 in the Kindle edition) Here, the z-score is calculated using the
\code{stats::qnorm()} function.
}
\details{
Also known as "Newcombe's method".
}
\examples{
proportion_difference(
  true_calls_1 = 99,
  false_calls_1 = 1,
  true_calls_2 = 80,
  false_calls_2 = 1,
  interval = 0.95
  )
}
