% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_difference.R
\name{proportion_difference_fraction}
\alias{proportion_difference_fraction}
\title{Calculate the confidence interval of a difference in proportions (e.g.
sensitivity or specificity)}
\usage{
proportion_difference_fraction(
  numerator_1,
  denominator_1,
  numerator_2,
  denominator_2,
  interval = 0.95,
  digits = 0.1
)
}
\arguments{
\item{numerator_1}{The numerator of the first proportion of interest.
Typically the number of true calls for a proportion such as PPA or NPA.}

\item{denominator_1}{The total number of calls for the first proportion of
interest.}

\item{numerator_2}{The numerator of the second proportion of interest.
Typically the number of true calls for a proportion such as PPA or NPA.}

\item{denominator_2}{The total number of calls for the second proportion of
interest.}

\item{interval}{The confidence interval to calculate. Defaults to
\code{0.95} for 95\%.}

\item{digits}{The number of digits to round to. Defaults to \code{0.1}.}
}
\value{
A list containing the estimated difference, the confidence interval,
a boolean attribute indicating whether or not a difference was detected at
the specified confidence (\code{difference_detected}), and various strings
and formatted percents that can be used to convey the information.
}
\description{
This is a wrapper around \code{proportion_difference()} that permits use of a
numerator and denominator rather than true calls and false calls.
}
\examples{
proportion_difference_fraction(
   numerator_1 = 99,
   denominator_1 = 100,
   numerator_2 = 79,
   denominator_2 = 80,
   interval = 0.95
)
}
