% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_evaluation_sheet.R
\name{read_evaluation}
\alias{read_evaluation}
\title{Read Evaluation}
\usage{
read_evaluation(filepath)
}
\arguments{
\item{filepath}{The filepath of the evaluation data to be read in from excel}
}
\value{
The panel data, equivalent to what was generated by
\code{build_evaluation_sheet()}, read in from excel.
}
\description{
Read Evaluation
}
