% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_panel_sheet.R
\name{read_panel}
\alias{read_panel}
\title{Read Panel}
\usage{
read_panel(filepath)
}
\arguments{
\item{filepath}{The filepath of the panel data to be read in from excel}
}
\value{
The panel data, equivalent to what was generated by
\code{build_panel_sheet()}, read in from excel.
}
\description{
Read Panel
}
