% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_pos_neg}
\alias{replace_pos_neg}
\title{Replace Positive and Negative}
\usage{
replace_pos_neg(string)
}
\arguments{
\item{string}{The strign to modify}
}
\value{
A string with words replaced by symbols.
}
\description{
A helper function to replace " Positive" with "+" and " Negative" with "-".
}
