% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twotest.R
\name{risk_analysis_series}
\alias{risk_analysis_series}
\title{Risk analysis for two tests in series}
\usage{
risk_analysis_series(n, prev, se1, sp1, se2, sp2)
}
\arguments{
\item{n}{The population size}

\item{prev}{The prevalence of the condition of interest}

\item{se1}{The sensitivity estimate for test 1.}

\item{sp1}{The specificity estimate for test 1.}

\item{se2}{The sensitivity estimate for test 2.}

\item{sp2}{The specificity estimate for test 2.}
}
\value{
A list of population and combined test statistics.
}
\description{
Risk analysis for two tests in series
}
