% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_evaluation.R
\name{score_evaluation}
\alias{score_evaluation}
\title{Score Evaluation}
\usage{
score_evaluation(panel_data, evaluation_data)
}
\arguments{
\item{panel_data}{The panel data associated with the evaluation, as read in
by \code{read_panel()}.}

\item{evaluation_data}{The evaluation data associated with the evaluation, as
read in by \code{read_evaluation()}.}
}
\value{
A table of scored results with qualitative results from the
evaluation compared to ground truth in the panel.
}
\description{
Score Evaluation
}
