% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_agreement.R
\name{sensitivity_fraction}
\alias{sensitivity_fraction}
\title{Calculate an agreement proportion for Sensitivity from a Fraction}
\usage{
sensitivity_fraction(
  numerator,
  denominator,
  digits = 0.1,
  interval = 0.95,
  ...
)
}
\arguments{
\item{numerator}{The numerator. The number of true positive calls.}

\item{denominator}{The total number of calls.}

\item{digits}{The number of digits to round to. Defaults to \code{0.1}.}

\item{interval}{The confidence interval to calculate. Defaults to
\code{0.95} for 95\%.}

\item{...}{Parameters passed to \code{agreement()}.}
}
\value{
A formatted agreement proportion.
}
\description{
This is a wrapper around \code{agreement()} that permits use of a
numerator and denominator rather than true calls and false calls.
}
\examples{
sensitivity_fraction(100, 100)
sensitivity_fraction(100, 100, 0.01, 0.99)
}
