% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_by_two.R
\name{two_by_two}
\alias{two_by_two}
\title{Two by Two}
\usage{
two_by_two(panel_data, evaluation_data)
}
\arguments{
\item{panel_data}{The result of \code{read_panel()}.}

\item{evaluation_data}{The result of \code{read_evaluation()}.}
}
\value{
A two-by-two table of outcomes
}
\description{
Generate the "two by two" table of results observed in the evaluation vs.
those expected based on the ground truth in the panel.
}
\examples{
two_by_two(dxr::example_panel_1, dxr::example_evaluation_1)
}
