% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twotest.R
\name{two_test_series}
\alias{two_test_series}
\title{Two tests in series}
\usage{
two_test_series(se1, sp1, se2, sp2, prev)
}
\arguments{
\item{se1}{The sensitivity estimate for test 1.}

\item{sp1}{The specificity estimate for test 1.}

\item{se2}{The sensitivity estimate for test 2.}

\item{sp2}{The specificity estimate for test 2.}

\item{prev}{The estimated prevalence of the condition of interest.}
}
\value{
A list containing the sensitivity estimate (\code{se_estimate}),
specificity estimate (\code{sp_estimate}), positive predictive value
estimate (\code{ppv_estimate}), and negative predictive value estimate
(\code{npv_estimate}) of the two tests used together.
}
\description{
Calculate sensitivity, specificity, and predictive value of two tests used in
series.
}
