% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validation_string}
\alias{validation_string}
\title{Validation String}
\usage{
validation_string(input_validation, column_vector)
}
\arguments{
\item{input_validation}{The table used to validate inputs in
\code{panel_sheet_excel_method()} and
\code{evaluation_sheet_excel_method()}.}

\item{column_vector}{The vector used to create the relevant column of
\code{input_validation}.}
}
\value{
A string describing the location of the list of valid inputs in the
excel file.
}
\description{
Build a string describing where the valid values are in Excel. Do this by
extracting the column number of \code{column_vector} in
\code{input_validation} and converting it to a letter per Excel's column
naming convention. Then use the length of the vector used to form that column
to set the relevant rows.
}
