% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_panel_sheet.R
\name{write_panel_sheet}
\alias{write_panel_sheet}
\title{Write Panel Sheet}
\usage{
write_panel_sheet(panel_sheet_data, filepath, method = "excel")
}
\arguments{
\item{panel_sheet_data}{The panel data, which are the output of
\code{dxr::build_panel_sheet()}.}

\item{filepath}{The filepath to write the panel sheet to.}

\item{method}{File format to output to. Currently, the only option is the
default, \code{"excel"}.}
}
\value{
An panel sheet template file, written to disk.
}
\description{
Generate a template sample panel sheet to record ground truth for an
evaluation or series of evaluations.
}
