% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_agreement.R
\name{agreement}
\alias{agreement}
\title{Calculate an agreement proportion}
\usage{
agreement(
  true_calls,
  false_calls,
  digits = 0.1,
  interval = 0.95,
  format_big = FALSE
)
}
\arguments{
\item{true_calls}{The number of correct calls (true positives or true
negatives).}

\item{false_calls}{The number of incorrect calls (false positives or false
negatives).}

\item{digits}{The number of digits to round to. Defaults to \code{0.1}.}

\item{interval}{The confidence interval to calculate. Defaults to \code{0.95}
for 95\%.}

\item{format_big}{A boolean indicating whether or not to format large numbers
in string outputs to include commas as a big mark and to avoid scientific
notation. Defaults to \code{FALSE}.}
}
\value{
A formatted agreement proportion.
}
\description{
Calculate an agreement proportion
}
\examples{
agreement(100, 0)
agreement(100, 0, 0.01, 0.99)
}
