% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_panel_sheet.R
\name{build_panel_sheet}
\alias{build_panel_sheet}
\title{Build Panel Sheet}
\usage{
build_panel_sheet(
  panel_name,
  panel_description = NA_character_,
  n_samples,
  sample_groups = c("Positive", "Negative"),
  sample_matrices,
  analytes,
  targets,
  qualitative_outcomes = c("Positive", "Negative"),
  qualitative_comparators,
  semiquantitative_outcomes = NA_character_,
  semiquantitative_comparators,
  quantitative_units = NA_character_,
  quantitative_comparators
)
}
\arguments{
\item{panel_name}{The name of the panel being used for the evaluation. Must
be a character string (not a multi-level vector) and must not be \code{NA}.}

\item{panel_description}{An optional sentence or paragraph describing the
panel. Defaults to \code{NA}. If provided, it must be a string rather than
a multi-level vector.}

\item{n_samples}{The number of samples in the panel. Must be numeric and
greater than 0. If not provided as an integer, it will be converted into
one.}

\item{sample_groups}{The descriptive terms used to describe groups of
samples, provided as a character vector. Defaults to \code{c("Positive",
  "Negative")} for evaluations of qualitative tests, but additional levels
can be added for, for example, potentially cross-reactive negative
subgroups. This must be a character vector with at least one level.}

\item{sample_matrices}{The matrix or matrices in which the samples are
provided. For example, for a serology assay panel, this could be
\code{"Serum"} for serum-only,  \code{"EDTA Plasma"} for plasma only, or
\code{c("Serum", "EDTA Plasma")} for evaluations involving multiple sample
types. This must be provided with at least one level (not \code{NA}).}

\item{analytes}{The analyte or analytes for which the panel has been
characterized and truth has been established. For example, for a serology
assay panel, this could be \code{"Pan-Ig"} or \code{c("IgM", "IgG",
  "Pan-Ig")}. This must be provided with at least one level (not \code{NA}).}

\item{targets}{The target or targets for which the panel has been
characterized and truth has been established. For example, for a SARS-CoV-2
assay, this could be \code{"Spike"} or \code{"RBD"}. Multiple targets can
also be used, such as \code{c("Spike", "Nucleocapsid")}. This must be
provided with at least one level (not \code{NA}).}

\item{qualitative_outcomes}{The valid qualitative outcomes associated with
the panel. Typically, this should be \code{c("Positive", "Negative")}. Note
that while the assay being evaluated may also provide an "Equivocal" result
level, this level should not typically be included in the panel sheet,
where ground truth is expected to have been determined for each sample
present. This must be provided with at least one level (not \code{NA}).}

\item{qualitative_comparators}{The qualitative comparator methods used to
determine qualitative ground truth for the panel. For example, if an
FDA-authorized PCR assay is used to determine ground truth--positive or
negative--for an evaluation of an antigen assay, this can be provided as
\code{"FDA-authorized PCR assay"} (or, ideally, the specific assay used,
sample matrix, etc.). If different methods were used to characterize
different samples, provide each method as an element of a character vector.
For example, \code{c("FDA-authorized PCR Assay 1", "FDA-authorized PCR
  Assay 2", "Both FDA-authorized PCR Assay 1 and 2")}.}

\item{semiquantitative_outcomes}{The valid semi-quantitative outcomes
associated with the panel. Defaults to \code{NA}, indicating that no
semi-quantitative ground truth has been established for the sample panel.
If titers or other information are available, the relevant levels must be
provided as a character vector. For example: \code{c("0", "100", "400",
  "1600", "6400")} (where \code{0} corresponds to a sample with a negative
qualitative result).}

\item{semiquantitative_comparators}{The semi-quantitative comparator methods
used to determine qualitative ground truth for the panel. Works similarly
to \code{qualitative_comparators} above. If
\code{semiqquantitative_outcomes} are provide (not all \code{NA}), then
this must be provided as well. If \code{semiqquantitative_outcomes} are not
provide (\code{all(is.na(semiqquantitative_outcomes))}), then this must be
NA as well.}

\item{quantitative_units}{If quantitative outcomes have been established for
the panel, this is a character string describing the units of those
quantitative results. Defaults to \code{NA}, indicating that no
quantitative ground truth has been established for the sample panel. If the
results are unit-less, \code{"Unit-less"} can be used.}

\item{quantitative_comparators}{The quantitative comparator methods used to
determine qualitative ground truth for the panel. Works similarly to
\code{qualitative_comparators} above. If \code{quantitative_units} is
provide (not \code{NA}), then this must be provided as well. If
\code{quantitative_units} is not provide
(\code{is.na(quantitative_units)}), then this must be NA as well.}
}
\value{
Returns \code{TRUE} if a sample panel sheet is created successfully.
}
\description{
A function to build an excel file with fill-in-the-blank cells for common
information needed about a sample panel or ground-truth.
}
\examples{
build_panel_sheet(
  panel_name = "Example Panel",
  panel_description = NA_character_,
  n_samples = 110L,
  sample_groups = c("Positive", "Negative"),
  sample_matrices = c("Serum", "Plasma"),
  analytes = c("IgM", "IgG", "Pan-Ig"),
  targets = c("Spike", "Nucleocapsid"),
  qualitative_outcomes = c("Positive", "Negative"),
  qualitative_comparators =
    c(
      "CDC Spike Antigen Assay and NCI Implementation of MS RBD Assay",
      "Collected Pre-2020"
    ),
  semiquantitative_outcomes = c("0", "100", "400", "1600", "6400"),
  semiquantitative_comparators = "CDC Spike Antigen Assay",
  quantitative_units = NA_character_,
  quantitative_comparators = NA_character_
)
}
