\docType{package}
\name{dxR-package}
\alias{dxR}
\alias{dxR-package}
\title{DNAnexus R Client Library}
\description{
  dxR is an R extension containing API wrapper functions
  for interacting with the new DNAnexus platform.
}
\details{
  \tabular{ll}{ Package: \tab dxR\cr Type: \tab Package\cr
  Version: \tab 0.125.0\cr License: \tab Apache License (==
  2.0)\cr }
}
\author{
  Katherine Lai

  Maintainer: Katherine Lai <klai@dnanexus.com>
}

